Move IImageCache to own file
